Switch Back to Using the .drone.yml For Now
Switched back to using the .drone.yml instead of .drone.star because the
starlark script wasn't consistently detecting pushes to the pre-release
tag. We will re-visit the starlark file and making that work later. For
now, we have more important things to focus on to get Lucky's initial
release out.