Reorganize Code To Support Building Without Daemon
This does a wide reorganization of the code to allow you to build Lucky
without the "daemon" feature for use by Lucky charm developers.

This also updates the CI builds for Windows and Mac to build without the
daemon feature. Linux is currently still building with the "daemon"
feature because the Daemon will technically work on Linux, even though a
charm developer will not manually run the daemon themselves.